Split workgroup unit test headers
