+++ test/unit/workgroup/tests/test-workgroup-WorkStorage-Iterator.hpp
@@ -1,0 +1,102 @@
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+// Copyright (c) 2016-21, Lawrence Livermore National Security, LLC
+// and RAJA project contributors. See the RAJA/COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+
+///
+/// Header file containing tests for RAJA workgroup constructors.
+///
+
+#ifndef __TEST_WORKGROUP_WORKSTORAGEITERATOR__
+#define __TEST_WORKGROUP_WORKSTORAGEITERATOR__
+
+#include "RAJA_test-workgroup.hpp"
+#include "test-util-workgroup-WorkStorage.hpp"
+
+#include <random>
+#include <array>
+#include <cstddef>
+
+
+template <typename StoragePolicy,
+          typename Allocator
+          >
+void testWorkGroupWorkStorageIterator()
+{
+  bool success = true;
+
+  using Vtable_type = RAJA::detail::Vtable<void, void*, bool*, bool*>;
+  using WorkStorage_type = RAJA::detail::WorkStorage<
+                                                      StoragePolicy,
+                                                      Allocator,
+                                                      Vtable_type
+                                                    >;
+
+  using callable = TestCallable<int>;
+
+  const Vtable_type* vtable = RAJA::detail::get_Vtable<
+      callable, Vtable_type>(RAJA::seq_work{});
+
+  {
+    WorkStorage_type container(Allocator{});
+
+    ASSERT_EQ(container.end()-container.begin(), (std::ptrdiff_t)0);
+    ASSERT_FALSE(container.begin() < container.end());
+    ASSERT_FALSE(container.begin() > container.end());
+    ASSERT_TRUE(container.begin() == container.end());
+    ASSERT_FALSE(container.begin() != container.end());
+    ASSERT_TRUE(container.begin() <= container.end());
+    ASSERT_TRUE(container.begin() >= container.end());
+
+    container.template emplace<callable>(vtable, callable{-1});
+
+    ASSERT_EQ(container.end()-container.begin(), (std::ptrdiff_t)1);
+    ASSERT_TRUE(container.begin() < container.end());
+    ASSERT_FALSE(container.begin() > container.end());
+    ASSERT_FALSE(container.begin() == container.end());
+    ASSERT_TRUE(container.begin() != container.end());
+    ASSERT_TRUE(container.begin() <= container.end());
+    ASSERT_FALSE(container.begin() >= container.end());
+
+    {
+      auto iter = container.begin();
+
+      ASSERT_EQ(&*iter, &iter[0]);
+
+      ASSERT_EQ(iter++, container.begin());
+      ASSERT_EQ(iter--, container.end());
+      ASSERT_EQ(++iter, container.end());
+      ASSERT_EQ(--iter, container.begin());
+
+      ASSERT_EQ(iter+1, container.end());
+      ASSERT_EQ(1+iter, container.end());
+      ASSERT_EQ(++iter, container.end());
+      ASSERT_EQ(iter-1, container.begin());
+      ASSERT_EQ(iter-=1, container.begin());
+      ASSERT_EQ(iter+=1, container.end());
+    }
+  }
+
+  ASSERT_TRUE(success);
+}
+
+
+template <typename T>
+class WorkGroupBasicWorkStorageIteratorUnitTest : public ::testing::Test
+{
+};
+
+TYPED_TEST_SUITE_P(WorkGroupBasicWorkStorageIteratorUnitTest);
+
+
+TYPED_TEST_P(WorkGroupBasicWorkStorageIteratorUnitTest, BasicWorkGroupWorkStorageIterator)
+{
+  using StoragePolicy = typename camp::at<TypeParam, camp::num<0>>::type;
+  using Allocator = typename camp::at<TypeParam, camp::num<1>>::type;
+
+  testWorkGroupWorkStorageIterator< StoragePolicy, Allocator >();
+}
+
+#endif  //__TEST_WORKGROUP_WORKSTORAGEITERATOR__
